requirements.txt
azure-cosmos
streamlit
